configurações iniciais dos arquivos

+++ backend/src/main/java/sptech/horticontrol/repository/ProdutoRepository.java
@@ -3,6 +3,5 @@ package sptech.horticontrol.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sptech.horticontrol.entity.Produto;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
-
+public interface ProdutoRepository extends JpaRepository<Produto, Long> {
 }
